Move decodeBech32Pubkey to pubkey.ts
