+++ packages/sdk/types/address.d.ts
@@ -1,7 +1,6 @@
-import { Bech32PubKey, PubKey } from "./types";
+import { PubKey } from "./types";
 export declare type CosmosAddressBech32Prefix = "cosmos" | "cosmosvalcons" | "cosmosvaloper";
 export declare type CosmosPubkeyBech32Prefix = "cosmospub" | "cosmosvalconspub" | "cosmosvaloperpub";
 export declare type CosmosBech32Prefix = CosmosAddressBech32Prefix | CosmosPubkeyBech32Prefix;
-export declare function decodeBech32Pubkey(bech: Bech32PubKey): PubKey;
 export declare function isValidAddress(address: string): boolean;
 export declare function encodeAddress(pubkey: PubKey, prefix: string): string;
