+++ packages/sdk/types/index.d.ts
@@ -1,9 +1,9 @@
 import * as logs from "./logs";
 import * as types from "./types";
 export { logs, types };
-export { CosmosBech32Prefix, decodeBech32Pubkey, encodeAddress, isValidAddress } from "./address";
+export { CosmosBech32Prefix, encodeAddress, isValidAddress } from "./address";
 export { unmarshalTx } from "./decoding";
 export { encodeSecp256k1Signature, makeSignBytes, marshalTx } from "./encoding";
 export { RestClient, TxsResponse } from "./restclient";
 export { makeCosmoshubPath, Pen, PrehashType, Secp256k1Pen } from "./pen";
-export { encodeSecp256k1Pubkey } from "./pubkey";
+export { decodeBech32Pubkey, encodeSecp256k1Pubkey } from "./pubkey";
